refactor(settings): SettingsContract 및 ViewModel 구조 개선
- SettingsContract 인터페이스 추가
- 기존 MviViewModel에서 BaseViewModel로 변경
- 이벤트 및 상태 관리 방식 개선
