+++ settings/presentation/src/main/java/io/soma/cryptobook/settings/presentation/SettingsContract.kt
@@ -3,21 +3,23 @@ package io.soma.cryptobook.settings.presentation
 import io.soma.cryptobook.core.domain.model.CurrencyUnit
 import io.soma.cryptobook.core.domain.model.Language
 import io.soma.cryptobook.core.domain.model.UserData
-import io.soma.cryptobook.core.presentation.Event
-import io.soma.cryptobook.core.presentation.SideEffect
-import io.soma.cryptobook.core.presentation.UiState
+import io.soma.cryptobook.core.presentation.mvi.UnidirectionalViewModel
 
-data class SettingsUiState(
-    val userData: UserData? = null,
-    val isLoading: Boolean = true,
-) : UiState
+interface SettingsContract {
+    interface ViewModel : UnidirectionalViewModel<State, Event, Effect>
 
-sealed interface SettingsSideEffect : SideEffect
+    data class State(
+        val userData: UserData? = null,
+        val isLoading: Boolean = true,
+    )
 
-sealed interface SettingsEvent : Event {
-    data class SetLanguage(val language: Language) : SettingsEvent
-    data class SetCurrencyUnit(val currencyUnit: CurrencyUnit) : SettingsEvent
-    data object NavigateToHome : SettingsEvent
-    data object ShowLoadingMessage : SettingsEvent
-    data object ShowSnackbarMessage : SettingsEvent
+    sealed interface Event {
+        data class SetLanguage(val language: Language) : Event
+        data class SetCurrencyUnit(val currencyUnit: CurrencyUnit) : Event
+        data object NavigateToHome : Event
+        data object ShowLoadingMessage : Event
+        data object ShowSnackbarMessage : Event
+    }
+
+    sealed interface Effect
 }
